MMI-244 Fixed CBRA report issues
- fixed the time zone issues that the publish on time
in db has time zone info, updated the view to convert it
to local time
- updated staff summary time to show in hh:mm:ss format

+++ libs/net/reports/CBRAReport.cs
@@ -200,7 +200,7 @@ public class CBRAReport
 
             cellB = row.CreateCell(1);
             cellB.CellStyle = valueStyle;
-            cellB.SetCellValue((double)timeTrack.CbraHours);
+            cellB.SetCellValue(timeTrack.CbraHours);
         }
         return startRowIndex;
     }
